+++ CHANGES.md
@@ -1,6 +1,11 @@
 CHANGES
 =======
 
+## Version 0.2 (2013-02-03T16:39:59Z)
+
+ * add option -c for Commit Calendar
+ * add option -U for not using Unicode
+
 ## Version 0.1 (2013-01-31T09:10:33Z)
 
  * add option -C for no color output